fix(multipart): use alternative method
Without the use of the [`multipart/alternative`][alternative] content
type some email clients (Outlook Web Access for one) show only the
plain text part of the email. The HTML part, then, is received as an
attachment (called `ATT00001.htm` in the case of OWA) and is not
displayed.

This patch uses `EmailBuilder.alternative()` in place of separate calls
to `EmailBuilder.text()` and `EmailBuilder.html()`, which ensures that
the `Content-Type: multipart/alternative` header is added.

[alternative]: https://en.wikipedia.org/wiki/MIME#Alternative